requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-tailwind==4.4.2
Pillow==11.0.0
pytest==8.4.2
pytest-django==4.11.1
